Grunt express server integration tests for GET actions. Refs #1.
Register a grunt task to put a json config file in the working directory for the express area server to use.

Use __dirname global to make path to env.json for the server settings. Used when working directory isn't the same as the app (i.e. Grunt).

Tidy up json-local provider to use one parseJson function to read the file and parse into json.